requirements.txt
APScheduler==3.6.3
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorclass==2.2.2
docopt==0.6.2
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pip-upgrader==1.4.15
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
terminaltables==3.1.10
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
zipp==3.7.0
